useRecoilTransactionObserver hook (#309)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/309

Initial `useRecoilTransactionObserver()` hook based on [this proposal](https://fb.quip.com/UwKaAGjK0L0E)

```
function useRecoilTransactionObserver(
  ({
    snapshot: Snapshot,
    previousSnapshot: Snapshot,
  }) => void,
): void
```

This adds support for selectors that the old `useTransactionObservation()` hook doesn't.  It contains both `snapshot` and `previousSnapshot`, but the snapshots don't yet have a mechanism for iterating the set of atoms/selectors or dirty atoms.  So, keep the old hook for now and migrate users over as we add that functionality.

Also mark `useTransactionSubscription()` as DEPRECATED, we will need to migrate the prototype dev tools support over.

Reviewed By: habond, csantos42

Differential Revision: D21737837

fbshipit-source-id: a9dd99431609e753898edfcb604b38a3508fda52