useGotoRecoilSnapshot() hook (#310)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/310

Add `useGotoRecoilSnapshot()` hook as per [this proposal](https://fb.quip.com/UwKaAGjK0L0E)

```
function useGotoRecoilSnapshot(): Snapshot => void
```

Change current store to use the state in the provided `Snapshot`.  If any atom values changed, then fire subscriptions to update any currently subscribed components for only those updated atoms.

Note that `nodeToComponentSubscriptions` is maintained as the state is updated to match the `Snapshot`.  This won't be necessary to explicitly copy when that state is migrated from the `TreeState` to the `StoreState` for CM.

I think the old implementation had a bug where it would only change atom values that were explicitly "updated" in the snapshot, making it not really a snapshot but a queue of state changes.

Next we will migrate `<Link>` and remove the old deprecated version.

Reviewed By: habond, csantos42

Differential Revision: D21739680

fbshipit-source-id: 5da9b3b13a10af208045413e5a73c65b55bb8c2b