Faster hooks
Summary:
Currently the most common Recoil hooks are implemented in terms of the more complex and deprecated `useInterface` hook, in order to reduce code redundancy. This imposes a large performance penalty on the common case.

In particular, the slow speed of these hooks is blocking Proton from using Recoil in a large number of components.

This diff reduces the time taken by `useRecoilState` from an average of 17 microseconds to 2 microseconds, as compared with around 1 microsecond for `useState` by itself.

The downside is some code semi-duplication between these hooks and `useInterface`, but hopefully this condition is short-lived.

Reviewed By: drarmstr

Differential Revision: D22543064

fbshipit-source-id: 52d50c2ad8f8ec2ec12d66a9de12073905ad3f81