Memoize atom values in each component
Summary:
Restores a long-lost optimization where the value of each atom is memoized within each component. Now, when rendering the component, we no longer have to call `getRecoilValueAsLoadable` for an atom unless that specific atom has been updated.

Note that the implementation depends on the fact that the `TreeState` returned by `useMutableSource` is the same until the corresponding subscription fires.

Besides optimization, this is important because it means that **atoms are never read under an outdated TreeState anymore**.

Reviewed By: drarmstr

Differential Revision: D23642523

fbshipit-source-id: 4c83eb3949f0125d109b4a85173273ac13f685e7