React Native bundle (#114)
Summary:
Fixes https://github.com/facebookexperimental/Recoil/issues/99

**Additions:**
- Add the ability to set/get a batch function, which defaults to `ReactBatchedUpdates`.
- Add `ReactBatchedUpdates` files for Web and Native that are imported to set the batch function to the renderer-specific function.
- Add a `native` target to Rollup, which resolves `ReactBatchedUpdates`.

**Result:**
This gives us out of the box support for React DOM and React Native, with the library user not needing to make any changes.

**Notes:**
I haven't changed the use of `unstable_batchedUpdates` in the tests. (In fact I haven't been able to get the suite to pass on `master`, which I'll have another go at later, and file an issue if I'm still stuck.) But it'd be great if the maintainers or anyone with more contexts on the tests could weigh in here. It seems like we should use `unstable_batchedUpdates` from `react-test-renderer` here.

In future, we could add a separate entry point (e.g. `'recoil/alternate-renderers'`) that doesn't set a batch function and doesn't depend on `react-dom` or `react-native`. We should be able to do this really easily because the infrastructure for setting the batching function to anything is available (and already exported in this PR).

**Prior art for this PR:**
- React Redux: [Alternate renderers entry point](https://github.com/reduxjs/react-redux/blob/master/src/alternate-renderers.js) and [get/set batch functions](https://github.com/reduxjs/react-redux/blob/master/src/utils/batch.js)
- dustinsoftware's [PR](https://github.com/facebookexperimental/Recoil/pull/108) with a similar implementation {emoji:1f64c}

Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/114

Reviewed By: drarmstr, davidmccabe

Differential Revision: D23733076

Pulled By: mondaychen

fbshipit-source-id: 948d45dd43b6e9ebac86ed6dd8433f3a1207c799