Memory management
Summary:
Optionally deletes the contents of atoms and selectors when they are no longer being used by components.

Atoms and selectors now have a 'retainedBy' option that specifies when they are deleted. Currently, the default is to be retained forever, as they are currently. A future change with code-mode should change this default.

Nodes that do not have root-lifetime retention can be retained in three ways:
* They are subscribed to by a component.
* They are retained by a component using a new `useRetain` hook, which doesn't subscribe the component but keeps the node for the duration that that component is mounted.
* They are a dependency of a selector that is retained (by the same possible means recursively).

The `useRetain` hook allows a top-level component to retain all of the state associated with some section of the app. But what if we'd rather not have a giant list of all the state in one place? What about code-splitting? To solve this, there is the new concept of a *retention zone*. A retention zone is just an opaque thing that can be referenced by both a node and a retaining component, allowing the component to retain the node without having a reference to it.
* A zone is created using the `retentionZone()` factory and may be passed to `useRetain` in place of a node. Every node within the zone is then retained.
* An node is part of a zone by providing that zone as its `retainedBy` option.
* An node can be part of any number of zones.

This adds `delete` to the `CacheImplementation` interface so that selectors can delete their cache entries. Existing implementations have been updated (arturg, twishas, tiansijie this is probably why you've been subscribed).

For families, the actual config for the individual atom or selector is deleted when that family entry is released. This is possible because the config can be lazily recreated when the same parameter is next used again. This means that the configs do not use memory.

Unanswered questions:

* Retention zones are based on reference equality of opaque objects, so how does that play with Fast Refresh? They could instead just be strings, maybe marked with an opaque type. What we don't want is a situation where we have a collection of all zones, because then they would also need memory-managed.

Reviewed By: drarmstr

Differential Revision: D24811801

fbshipit-source-id: 1d8e055a887eaf3e83c316e702fbafc5b373f3c3