Allow throwing non-errors
Summary:
Currently, if a selector throws a non Promise, the loadable will go to a 'hasValue' state. Throwing a non Promise should send the loadable into a 'hasError' state.

I've also updated the flow types to widen the `contents` type for Loadable from `Error` to `mixed` when in a 'hasError' state.

Reviewed By: drarmstr, csantos42

Differential Revision: D25622241

fbshipit-source-id: 898d13d00af38aac15decd3a8372ac78bd53b665