Update fbt and recoil to use the ReactDOMLegacy_DEPRECATED entry point (#1086)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1086

We're updating the internal entry point for the old React API to be ReactDOMLegacy_DEPRECATED. For the requires to keep working internally and externally I need to update the module mapping and the source to use these entry points.

See https://fb.workplace.com/groups/react.fyi/permalink/4411697998840538/ for context.

Reviewed By: rickhanlonii

Differential Revision: D29334366

fbshipit-source-id: 2ab2a118b2ea57094992dba29dc3ee49d28b2859