Implement useRecoilRefresher_UNSTABLE hook (#1294)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1294

It is often desirable to be able to re-run a selector in a Recoil application. There is currently no first class way of doing this. The suggested approach is to define an upstream atom (such as a number), which your selector depends on. You can then bump this atom when you need the selector to re-run.

This new approach adds a first class hook which can be used to force a selector to re-run.

Internally it simply clears the cache, and marks the value as updated.

Reviewed By: drarmstr

Differential Revision: D31579872

fbshipit-source-id: 5b862f6ec86e992283a169993640870d76460c91