Mode for Recoil to sync with query parameters (#1335)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1335

We have a variety of modes for how Recoil Sync should serialize atoms to the URL.  For example:

* `href` - User has complete control over creating the URI
* `anchor` - User can serialize all atoms to an anchor string
* `search` - User can serialize all atoms to the query search string

This diff updates and clarifies the following two modes (which are probably the most useful):

* `queryParams` - User can serialize individual atoms to sync with individual query parameters
* `queryParams` with a `param` - User can serialize all atoms to a single query parameter string.

This flexibility is helpful because sometimes we want to use a mode like `{part: 'queryParams', param: 'state'}` to serialize with something like Google Transit to persist complex user objects in a single string with compression/de-duping while other times we may want a mode like `{part: 'queryParams'}` that for some trivial JSON serialization of atoms to their own query parameters.  That may not support complex objects or user classes, but is convenient for others to read the URL, such as for Relay pre-fetching.  the `{part: 'href'}` is a sort of fallback mechanism to allow user complete freedom.

Note that with this flexibility the `serialize()`/`deserialize()` callback signatures change to just be `mixed => string` and `string => mixed` and depending on the mode used with either pass individual atom values to serialize or an object containing all atom values.  We don't want to just always have the callback deal with individual atoms at a time because that would remove the opportunity for de-duping optimizations such as available with Google Transit.

Reviewed By: geekster777

Differential Revision: D31873185

fbshipit-source-id: 1e8aad6c0d8d4717e5631153107fc9a5496f15f3