Support atoms with fallback selectors for Transit serialization (#1427)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1427

Support serializing atoms with selector defaults for Transit serialization for URL synchronization.  Note that we are only serializing the explicitly set state of the atom or the state that it should use the fallback selector, not the state of the fallback selector itself.

When an atom has a fallback selector it technicall creates a wrapper selector and the base atom retains the value of `DefaultValue` until it is explicitly set.  Thus, we need to support serialization of `DefaultValue` to handle this state.

Reviewed By: habond

Differential Revision: D32452527

fbshipit-source-id: 7b9aeacc31727d92b8d9e34af13a4ded1d82b658