Automate testing React Concurrent Mode with Recoil (#1490)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1490

Automate testing React Concurrent Mode with Recoil.

Driveby fix of some unit tests that weren't confirming the `expect()`'s passed and were affected by Strict Mode.

We now will test the cross product of all of the GK combinations with Strict Mode on and off and Concurrent Mode on and off.  Unfortunately this is a large combination of 13,862 tests taking 40s.  Hopefully we can reduce this number as we settle on a method for syncing with React state.

NOTE: OSS React `18.0.0-rc.0` tests are currently not passing with `createRoot()` AND `<StrictMode>`.  Other OSS users reported a similar issue that may only affect Jest testing..  Disable that testing combination for now.

Reviewed By: habond

Differential Revision: D32945418

fbshipit-source-id: dcea9e584c3170f1814b058b6a04fabbd1c4cdee