Test onSet() is called when atom initialized via snapshot (#1519)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1519

Test that `onSet()` is called and `setSelf()` works when atoms are initialized first via Snapshot and then used from hooks in a `<RecoilRoot>`.  This includes when initializing atoms via `initializeState` prop in `<RecoilRoot>` which uses a Snapshot.  Also make sure the "current" value provided in the updater form of `setSelf()` reflects the current initialized value.

Reviewed By: habond

Differential Revision: D33372676

fbshipit-source-id: 7de0b837847e43a6ef1cc5d383f948053850de09