Avoid releasing retainables if there are pending writes (#1589)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1589

Avoid releasing retainables at the end of the batch if there were writes while processing them.  This could happen, for example, if there are effects which monitor changes and then also write state.

This error is currently superfulous for the open source release where garbage collection is enabled but not used.

Addresses #1582

Reviewed By: davidmccabe

Differential Revision: D33965759

fbshipit-source-id: 904f6e1778f7ba864cea54f0e3f2641e995834c1