Improved error message for invalid async sets (#1784)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1784

The current environment strips error message from classes extending `Error`.  So, instead just use a base `Error` type for error messages to be visible when trying to:
* set atoms/selectors using an updater function when the current state is pending.
* set selector with a get of an atom/selector that is in a pending state.

Note that these errors may still be missed as the sets are usually called from async handlers which do not catch exceptions.  We could potentially maybe cause these kinds of errors to put the `<RecoilRoot>` in an error state in dev mode, but that seems too aggressive.  Adding a console error for now.

In the long-term we could add support for asynchronous sets.  However, that will require updating the semantics of the API.  See these RFCs:
* https://fb.quip.com/913nA3Su8Ybr
* https://fb.quip.com/TUAhAfJI1CfJ

Reviewed By: ryprice

Differential Revision: D35693623

fbshipit-source-id: 68540f00fce25a56ff4a973cf6e35552a6406b25