Add annotations to empty objects that look like indexers in html/shared/
Summary:
We are working on making the empty object literal `{}` have the type `{}` - i.e. exact empty object - rather than being unsealed.
Making this change exposes a variety of errors. We can prevent these errors by annotating what we want the type of the empty object to be.

Details:
- Codemod script: `./scripts/typedjs/flow/runner lti/annotate_empty_objects html/shared`
- Announcement: [post](https://fb.workplace.com/groups/flowlang/posts/903386663600331)
- Support group: [Flow Support](https://fb.workplace.com/groups/flow)

drop-conflicts

Reviewed By: SamChou19815

Differential Revision: D36391213

fbshipit-source-id: dd06feca602b85e94b5c26612a478dbbe41d680d