Make variables required (#1791)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1791

Make the interface slightly more straightforward by making `variables` required for `graphQLSelectorFamily()`.  Previously the default would be to pass through the family parameter as the variables.  But, this may not be obvious and another assumption might be that leaving `variables` off would use empty variables or might skip the query.

Reviewed By: habond

Differential Revision: D36390363

fbshipit-source-id: 1ad3793f651e5ec82326eeed59f27cf942f2ef5f