Enable recoil-sync unit tests in GitHub (#1841)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1841

Enable the `recoil-sync` unit tests in the GitHub CI actions.  Add the `transit-js` dependency as part of the merged `package.json` which is now only used for development now that we have added `package-for-release.json` for the actual release of the individual packages.  A few other minor fixes for the RecoilSync unit tests to work in open source environment.

Reviewed By: mondaychen

Differential Revision: D36957102

fbshipit-source-id: ff36359dcc3c717ccc75c38be8ffc49242835c9d