Resolve promise on server side rendering as endBatch is only called on useEffect (#2073)
Summary:
Reference Issue: [https://github.com/facebookexperimental/Recoil/issues/1960](https://github.com/facebookexperimental/Recoil/issues/1960)

The function sendEndOfBatchNotifications is only executed on useEffect as below:
```
useEffect(() => {
    // enqueueExecution runs this function immediately; it is only used to
    // manipulate the order of useEffects during tests, since React seems to
    // call useEffect in an unpredictable order sometimes.
    Queue.enqueueExecution('Batcher', () => {
      endBatch(storeRef.current);
    });
  });
```

This caused the SSR usage of async selector to get stuck in infinite loading.

With this PR, we make sure that when the original promise is resolved, we resolve the `handleLoadable` promise as well. Thus making it compatible with SSR rendering

Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/2073

Reviewed By: wd-fb

Differential Revision: D40948099

Pulled By: drarmstr

fbshipit-source-id: 0a2c89972fc8a841b8271718bf8f212153ab3ce5